better handling of undefined from render functions in view models
Side effect of this change is to render undefined values as empty text
nodes. Originally an idea of @joshski's to reduce diff/patch operations
when adding and removing elements from the dom. Instead of the list of
vdom elements having one more or less item, we have the same number of
items, but some are empty text nodes.