Pre populate example code blocks
As opposed to generate code blocks at runtime.

Because generating code blocks at runtime does not play nicely with
other docsify plugins. E.g., there is a "copy-paste" plugin that does
not have a chance to hook into the example block, because when
"copy-paste" plugin kicks in, there are no code blocks yet.